+++ src/app/app.module.ts
@@ -25,19 +25,19 @@ import { environment } from '../environments/environment';
 //pages
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { GoogleMapComponent } from './components/google-map/google-map.component';
-import { IndoorNavigationSideButtonsComponent } from './components/indoor-navigation-side-buttons/indoor-navigation-side-buttons.component';
-import { IndoorNavigationToolbarComponent } from './components/indoor-navigation-toolbar/indoor-navigation-toolbar.component';
-import { ModalDirectionsComponent } from './components/modal-directions/modal-directions.component';
-import { OutdoorNavigationSideButtonsComponent } from './components/outdoor-navigation-side-buttons/outdoor-navigation-side-buttons.component';
-import { OutdoorNavigationToolbarComponent } from './components/outdoor-navigation-toolbar/outdoor-navigation-toolbar.component';
-import { PoiPopoverComponent } from './components/poi-popover/poi-popover.component';
-import { SearchPopoverComponent } from './components/search-popover/search-popover.component';
+import { GoogleMapComponent } from './components/outdoor/google-map/google-map.component';
+import { IndoorNavigationSideButtonsComponent } from './components/indoor/indoor-navigation-side-buttons/indoor-navigation-side-buttons.component';
+import { IndoorNavigationToolbarComponent } from './components/indoor/indoor-navigation-toolbar/indoor-navigation-toolbar.component';
+import { ModalDirectionsComponent } from './components/outdoor/modal-directions/modal-directions.component';
+import { OutdoorNavigationSideButtonsComponent } from './components/outdoor/outdoor-navigation-side-buttons/outdoor-navigation-side-buttons.component';
+import { OutdoorNavigationToolbarComponent } from './components/outdoor/outdoor-navigation-toolbar/outdoor-navigation-toolbar.component';
+import { PoiPopoverComponent } from './components/popovers/poi-popover/poi-popover.component';
+import { SearchPopoverComponent } from './components/popovers/search-popover/search-popover.component';
 import { SettingsToolbarComponent } from './components/settings-toolbar/settings-toolbar.component';
-import { IndoorMapComponent } from './components/indoor-map/indoor-map.component';
+import { IndoorMapComponent } from './components/indoor/indoor-map/indoor-map.component';
 import { OutdoorViewPage } from './pages/outdoor-view/outdoor-view.page';
 import { IndoorViewPage } from './pages/indoor-view/indoor-view.page';
-import { TimeFooterComponent } from './components/time-footer/time-footer.component';
+import { TimeFooterComponent } from './components/popovers/time-footer/time-footer.component';
 
 //services
 import { GeolocationServices } from 'src/services/geolocationServices';
