+++ src/app/components/indoor/indoor-map/indoor-map.component.html
@@ -1,5 +1,5 @@
 <div class="indoor-map-wrapper">
-  <div *ngIf="building === 'jmsb'">
+  <div *ngIf="inputBuilding === 'jmsb'">
     <mb1-floor-plan></mb1-floor-plan>
   </div>
 </div>
