+++ src/app/components/indoor/indoor-map/indoor-map.component.scss
