+++ src/app/components/indoor/indoor-navigation-side-buttons/indoor-navigation-side-buttons.component.scss
