+++ src/app/components/indoor/indoor-navigation-toolbar/indoor-navigation-toolbar.component.spec.ts
