+++ src/app/components/outdoor/google-map/google-map.component.scss
