+++ src/app/components/outdoor/google-map/google-map.component.spec.ts
