+++ src/app/components/outdoor/modal-directions/modal-directions.component.scss
