+++ src/app/components/outdoor/outdoor-navigation-side-buttons/outdoor-navigation-side-buttons.component.scss
