+++ src/app/components/outdoor/outdoor-navigation-toolbar/outdoor-navigation-toolbar.component.scss
