+++ src/app/components/outdoor/outdoor-navigation-toolbar/outdoor-navigation-toolbar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Events } from '@ionic/angular';
 import { DirectionService } from 'src/services/direction.service';
-import { DataSharingService } from '../../../services/data-sharing.service';
+import { DataSharingService } from '../../../../services/data-sharing.service';
 
 @Component({
   selector: 'app-outdoor-navigation-toolbar',
@@ -36,7 +36,9 @@ export class OutdoorNavigationToolbarComponent implements OnInit {
     });
   }
 
-  ngOnInit() {}
+  ngOnInit() {
+    this.loc = '0';
+  }
 
   sendMessage(updatedMessage) {
     this.data.updateMessage(updatedMessage);
