+++ src/app/components/popovers/poi-popover/poi-popover.component.scss
