+++ src/app/components/popovers/search-popover/search-popover.component.scss
