+++ src/app/components/popovers/time-footer/time-footer.component.scss
