+++ src/app/pages/indoor-view/indoor-view.page.html
@@ -1,7 +1,7 @@
 <ion-header>
-  <app-indoor-navigation-toolbar></app-indoor-navigation-toolbar>
+  <app-indoor-navigation-toolbar [inputBuilding]='building'></app-indoor-navigation-toolbar>
 </ion-header>
 <ion-content>
   <app-indoor-navigation-side-buttons></app-indoor-navigation-side-buttons>
-  <app-indoor-map [building]='building'></app-indoor-map>
+  <app-indoor-map [inputBuilding]='building'></app-indoor-map>
 </ion-content>
